photo memory cache

+++ apps/backend/src/main/java/no/nav/data/team/resource/ResourceService.java
@@ -1,5 +1,7 @@
 package no.nav.data.team.resource;
 
+import com.github.benmanes.caffeine.cache.Cache;
+import com.github.benmanes.caffeine.cache.Caffeine;
 import io.prometheus.client.Gauge;
 import lombok.extern.slf4j.Slf4j;
 import no.nav.data.team.common.security.azure.AzureAdService;
@@ -24,7 +26,9 @@ public class ResourceService {
             .name("team_profile_picture_count_gauge").help("Number of profile pictures cached")
             .register();
 
-    private static final Duration PHOTO_CACHE_DURATION = Duration.ofDays(1);
+    private static final Duration PHOTO_DB_DURATION = Duration.ofDays(1);
+    private static final Duration PHOTO_MEM_DURATION = Duration.ofHours(1);
+    private final Cache<String, byte[]> photoCache;
 
     private final StorageService storage;
     private final ResourcePhotoRepository resourcePhotoRepository;
@@ -34,10 +38,17 @@ public class ResourceService {
         this.storage = storage;
         this.resourcePhotoRepository = resourcePhotoRepository;
         this.azureAdService = azureAdService;
+        this.photoCache = Caffeine.newBuilder().recordStats()
+                .expireAfterWrite(PHOTO_MEM_DURATION)
+                .maximumSize(200).build();
     }
 
     @Transactional
-    public ResourcePhoto getPhoto(String ident, boolean forceUpdate) {
+    public byte[] getPhoto(String ident, boolean forceUpdate) {
+        var cached = photoCache.getIfPresent(ident);
+        if (!forceUpdate && cached != null) {
+            return cached;
+        }
         List<GenericStorage> photoStorage = resourcePhotoRepository.findByIdent(ident);
         if (forceUpdate) {
             photoStorage.forEach(photo -> storage.delete(photo.getId(), ResourcePhoto.class));
@@ -47,23 +58,28 @@ public class ResourceService {
         if (photoStorage.isEmpty()) {
             log.info("Get photo id={} calling graph", ident);
             var picture = azureAdService.lookupProfilePictureByNavIdent(ident);
-            return storage.save(ResourcePhoto.builder()
+            return getPhoto(storage.save(ResourcePhoto.builder()
                     .content(picture)
                     .ident(ident)
                     .missing(picture == null)
                     .build()
-            );
+            ));
         }
         if (photoStorage.size() > 1) {
             // Cleanup duplicates from race conditions
             photoStorage.subList(1, photoStorage.size()).forEach(ps -> storage.delete(ps.getId(), ResourcePhoto.class));
         }
-        return photoStorage.get(0).getDomainObjectData(ResourcePhoto.class);
+        return getPhoto(photoStorage.get(0).getDomainObjectData(ResourcePhoto.class));
+    }
+
+    private byte[] getPhoto(ResourcePhoto photo) {
+        photoCache.put(photo.getIdent(), photo.getContent());
+        return photo.getContent();
     }
 
     @Scheduled(initialDelayString = "PT1M", fixedRateString = "PT10M")
     public void cleanOld() {
-        log.debug("Deleted {} old photos", storage.deleteCreatedOlderThan(ResourcePhoto.class, LocalDateTime.now().minus(PHOTO_CACHE_DURATION)));
+        log.debug("Deleted {} old photos", storage.deleteCreatedOlderThan(ResourcePhoto.class, LocalDateTime.now().minus(PHOTO_DB_DURATION)));
     }
 
     @Scheduled(initialDelayString = "PT1M", fixedRateString = "PT1M")
